Added `name` property to Tokenizer
The `name` property can now be used to assist in integrating with HTML
forms.  Setting this will allow your tokenizer to submit the selected token
values as an array.

Also, rebuild the test and dist bundles, and update the example to set a
name and have a submit button

Partial fix for #9