bugfix: tokenizer shares state when no defaultSelection provided
React's `getDefaultProps` docs warn that due to caching, complex results
may be shared across components.  because getInitialState() uses the
inputs as-is, we ended up sharing state between tokenizers in broken ways.

Tokenizer was violating the recommendation, and the fix is to copy the
provided props in getInitialState.

This ended up breaking some of the tokenizer tests which assumed that
state would be shared across test runs (in spite of beforeEach).  I've
fixed that as well and have removed the explicit `clearSelection` calls
since this now happens implicitly by it being a separate test case.